Show reports published to NetworkTables
A list of reports is shown in a resizable side panel.  For any report in
the list that is checked, the video will, if possible, have an overlay
that draws the published values as shapes.

This currently works with line reports and blobs reports.  Contours,
numbers, and sizes are still TODO