Fix invalid magic number issue
When we read the image data, we have to use readFully.  Otherwise, if
GRIP is in the middle of sending the image, we could end up reading half
of it and interpretting the next few bytes as the header of the next
image, resulting in an "incorrect magic number" error.

In practice, this seems to happen when GRIP is publishing large,
high-quality images or is running on a slower machine.